backend/requirements.txt
# HevyCoach-AI Backend Dependencies
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
cryptography>=42.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development (optional)
httpx>=0.26.0  # For testing
